requirements.txt
crewai
crewai_tools
google-generativeai
python-dotenv